net: fec: fix ptp restart handling in AVB mode
The rebase of the commit c070eae9e9 ("net: fec: prevent enet counter
from being stopped on link status changes") on Linux 6.12.34 introduced
a bug in the FEC reset sequence as the code was recently [1] moved
to a separate fec_ctrl_reset) function. Moreover, multiple commits ([2], [3])
were added to handle, with a save/restore mechanism, the full reset of the MAC
on certain platforms causing the 1588 counter stop.
Since AVB support always relies on the fact that the FEC reset never disable
the 1588 counter (doing a MAC disable instead of a MAC reset), bypass all of the
ptp save/restore newly introduced code.

[1] 701118e6f6 ("net: fec: Refactor MAC reset to function")
[2] a1477dc87d ("net: fec: Restart PPS after link state change")
[3] c32fe1986f ("net: fec: Fix FEC_ECR_EN1588 being cleared on link-down")

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>