LF-16880 net: dsa: netc: correct max-sdu when it is default to 0
If the max-sdu value is not explicitly set when configuring tc-taprio,
the value defaults to 0. According to the tc-taprio manual page, the
value 0 means that the traffic class can send packets up to the port's
maximum MTU in size.

Currently, the driver does not check the value 0, it will set the max
SDU to 0 when max-sdu is not explicitly set. So all the packets cannot
be sent out from the switch port.

Signed-off-by: Wei Fang <wei.fang@nxp.com>