net: netc: align default queue maxsdu to port maxsdu
The queue maxsdu is also used for guardband of Qbv. It should be equal
to port maxsdu if the queue maxsdu is set to 0. The value of queue
maxsdu is 0 on default when using tc-taprio to set Qbv. If Qbv is not
configured, use default NETC_MAX_FRAME_LEN to set the queue maxsdu. This
will not affect the port SDU.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>